Cull children since the script does not mount parents
I don't thing this is entirely correct but I have no example at hand.
You should be able to create a filesystem on e.g. /dev/sde directly without
creating a partition /dev/sde1 first. So, technically this "parent" without
"children" should be mountable.

Since the script does not consider this. This commit filters for
elements/"parents" without "children".

For most cases this will keep the height of the output smaller not to run into a
curses crash. Not ideal and the curses crashes should get looked at, but for now
this will do to reduce crashes with small windows and large amount or block
devices – thanks snap.