Fix JSON decode error.
The JSON output from lsblk seems to have some issues that interfere with the `json.loads()` function. Specifically, the value 0B, which is expected to be a string.

By doing a direct replace on the lsblk output for `0B` on `menu.py`, the JSON is parsed and the menu appears.